Refactor prefill layer class names
